EthernetClientとWiFiClientをテンプレートで指定するようにした
